requirements.txt
discord.py
trueskill
ruamel.yaml
click
peewee
peewee-async
aiomysql
toolz
